webmidi/requestmidiaccess.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="UTF-8">
    <title>Web MIDI API Test: requestMIDIAccess()</title>
    <link rel="author" title="Ryoya KAWAI" href="mailto:ryoya.kawai@gmail.com">
    <link rel="help" href="http://webaudio.github.io/web-midi-api/#requestmidiaccess">
    <link rel="match" href="reference/REFERENCE-FILE-NAME">
    <meta name="assert" content="Return true when test success in requestMIDIAccess().">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>

    <style type="text/css">
    /* ADD STYLE BLOCK HERE (PREFERRABLE TO INLINE STYLES) */
    body { font-family:arial;}
    </style>
  </head>
  <body>
    <p>Success when requestMIDIAccess() exists and MIDIAccess Interface are correctly obtained .</p>
    <div id="log"></div>

    <!-- PAGE CONTENT -->
    <!--script type="text/javascript" src="webmidiapi_01_request.js"></script-->
    <script type="text/javascript">
    var MIDI;

    var test_requestMIDIAccess = async_test("requestMIDIAccess({sysex: flase})");
    function testWMA() {
        var ret=navigator.requestMIDIAccess({sysex:false}).then( successCallback, errorCallback );
        test(function(){
            assert_true(ret.constructor.name=="Promise");
        }, "Return type in {sysex:false}");
        function successCallback(access) {
            MIDI=access;
            test_requestMIDIAccess.step(function() {
                test(function(){
                    assert_true(MIDI.constructor.name=="MIDIAccess");
                }, "MIDIAccess Interface name test in {sysex:false}");
                test(function(){
                    assert_true(MIDI.inputs.constructor.name=="MIDIInputMap");
                    assert_true(typeof MIDI.onconnect=="object");
                    assert_true(typeof MIDI.ondisconnect=="object");
                    assert_true(MIDI.outputs.constructor.name=="MIDIOutputMap");
                    assert_false(MIDI.sysexEnabled);
                }, "MIDIAccess interface attribute test in {sysex:false}.");
                assert_true(true);
            });
            test_requestMIDIAccess.done();
        }
        function errorCallback() {
            test_requestMIDIAccess.step(function() {
                assert_false(true);
                test_requestMIDIAccess.done();
            });
        }
    }
    testWMA();

    MIDI="";
    var test_requestMIDIAccess02 = async_test("requestMIDIAccess({sysex: true}) - W3C Spec: Web MIDI API");
    function testWMA02() {
        var ret02=navigator.requestMIDIAccess({sysex:true}).then( successCallback, errorCallback );
        test(function(){
            assert_true(ret02.constructor.name=="Promise");
        }, "Return type test in sysex:true");
        function successCallback(access) {
            MIDI=access;
            test_requestMIDIAccess02.step(function() {
                test(function(){
                    assert_true(MIDI.constructor.name=="MIDIAccess");
                }, "MIDIAccess interface name test in {sysex:true}");
                test(function(){
                    assert_true(MIDI.inputs.constructor.name=="MIDIInputMap");
                    assert_true(typeof MIDI.onconnect=="object");
                    assert_true(typeof MIDI.ondisconnect=="object");
                    assert_true(MIDI.outputs.constructor.name=="MIDIOutputMap");
                    assert_true(MIDI.sysexEnabled);
                }, "MIDIAccess interface attribute test in {sysex:true}.");
                assert_true(true);
            });
            test_requestMIDIAccess02.done();
        }
        function errorCallback() {
            test_requestMIDIAccess02.step(function() {
                assert_false(true);
                test_requestMIDIAccess02.done();
            });
        }
    }
    testWMA02();
    </script>


  </body>
</html>
